add menu for editing sprays and pokos (#10)
Adds a new menu that edits the current spray and Poko count, unlocking
sprays if the demo flag is not set. Counters are validated such that
negative numbers are not allowed, spray counts above 99 are not allowed,
and Poko counts above 99,999 are not allowed.

The basic implementation is every menu option is a struct that inherits
from `MenuOption` that requires a `J2DTextBoxEx*` and
`CallBack_CounterRV*`, as well as any overrides for `up()`, `down()`,
`left()`, and `right()` to define what control stick directions do when
the option is being edited (default behavior is do nothing).

The menu doesn't look good, of course, but there are still many
decisions left to be made about how we want to organize all options
across all menus, as well as how the code is structured for each menu,
so I didn't put much effort into visuals.

Annotation convention I used: `@P2GZ` above entirely new
functions/structs/etc, `@P2GZ start` and `@P2GZ end` around multi-line
modifications to existing functions/structs/etc, and `@P2GZ` on the same
line for single-line modifications.

Also the work to actually enable the menu and updating the README was
stolen from HeartPiece's testing branch, thanks for doing that.