add squad editor (#12)
Lots of changes:
- Added a squad editor. The code is, like, kinda bad and there's lots of
copy-paste but I didn't want to prematurely refactor and move stuff
around until we make a few more screens and have a better sense of what
would actually be useful.
- Added custom asset patching. Needs to have the path added to the build
script. Maybe it's better if we just put all assets under `files` and
copy everything over the same way whether or not it's custom? That way
we don't need to specify custom assets in the build script, only the
directories that contain .szs archives.
- Added `--restart-dolphin` flag to build script. Currently restarting
is Windows-only and requires putting a shortcut to Dolphin.exe in the
root directory. Was thinking of adding a config file that you could put
your Dolphin installation directory in to make it cross-platform but
couldn't be bothered. Happy to add it or Maya can do it later :^)

Behavior of ending the day, saving, etc with Pikmin colors that are not
unlocked is not currently defined by P2GZ - however the vanilla game
handles those events with Pikmin colors that are not unlocked is what
happens. We can discuss ideal behavior later.

---------

Co-authored-by: Dean Southwood <deancsouthwood@gmail.com>