Re-implement code and asset build system (#30)
Re-implements relevant parts of the old build system, and (re)adds
dependencies on nodtools and cube to do iso and asset
unpacking/repacking. New or changed assets should be added to `files/`
along with the desired directory structure within the dol.

A .iso file for USA Pikmin 2 (GPVE01) now needs to be placed in the main
directory. Code and assets are built into an unpacked iso in `root/` by
running `python3 build.py`, and `root/sys/main.dol` can be run through
Dolphin to view changes.