New Game ID & Banner + redo build system slightly (#35)
Gives P2GZ a new unique internal game ID so it is no longer considered
Pikmin 2 to Dolphin/Native GameCubes (will not overwrite existing P2
saves). Also updated the internal game banner with new description text,
and gave it a unique icon for both its dolphin banner and save icon.

Additionally, reworked `build.py` slightly so that it now properly reads
any new assets in `files\` and adds them to the root folder. This allows
us to share non-code assets for upcoming features, and also allows the
builder to properly add it to the game. This is showcased with it
utilizing the new files I created for the banner and icons.