Add toggle to skip save prompts in caves (#65)
Does what it says on the tin - re-implements #25 (but without the
crashing). Currently under Settings, and skipping save prompts is off by
default - could change it so skipping is the default behaviour, but
figured for now we'd want saving on for testing - lemme know if we
should change this.

Probably scope for a future PR to update the timer to account for save
prompt skipping.