Skip boot-up screens (#66)
Skips epilepsy warning, Nintendo logo and Dolby logo to load straight
into the title screen. Also makes bootSection.cpp equivalent in the
process so it can be edited.

This is currently very conservative about what's skipped, since there's
some loading that seems to occur during the boot-up sequence and I don't
want to mess that up. So there is a bit of a black-screen delay before
the title screen, but far shorter than with the boot-up screens.