Fix: set mold cutscene to played for all PoD presets (#136)
For some untraceable reason, the area in front of the spiderwort plants
where the mold would spawn in Awakening Wood is triggering the "discover
spiderwort mold" cutscene, even when the mold hasn't spawned yet, such
as on Day 3. This fixes that (and fixes #135) by setting the cutscene to
"played" in all PoD presets, since we'd never encounter it in a run
anyway.

Down the track, we may have to manually turn it back on for some AT
presets (specifically enter SR), but we can deal with that later when we
have AT presets to think about.