Make itemBridge.cpp equivalent (#137)
Fixes some small errors in the decompilation of itemBridge.cpp to make
it functionally equivalent. Two functions are non-matching still, but
I've gone through them carefully and I'm confident that the float
regswaps are equivalent now with these small changes. Have also
confirmed with a play test, and bridges behave as expected in terms of
building, collision and interactions with objects. Hopefully this
facilitates some bridge-related features in the future.