Automatically center version print text (#138)
Automatically centers the version text (horizontally) that prints on the
title screen, so we don't have to experiment with placements in the
future as version strings change lengths. I considered making it an
inline, but handling the variadic stuff made it messy, so decided
against it - might be worth pursuing in the future if we're really keen
to do a lot more centered text printing.