Fix: Customise every enemy's debug info height (#134)
Fixes #131 and extends it by customising the debug info heights for all
enemies in the game. Some aren't changed because they look good as-is,
but others have been modified by a lot to be visible through each
enemy's main animations. Some small adjustments could be done down the
line based on feedback, but this at least makes them all legible and not
clipped horribly inside bodies.