Fix: make timer menu settings sync properly (#133)
Fixes #128 by creating a few new timer functions which update menu
settings when any menu is open (except the timer submenu). This means
the timer settings displayed in the submenu are synced, even when the
timer options are changed by gameplay (such as entering or leaving a
cave).

Also adds a couple helper inlines for checking (by title) if we have a
given menu open, or if we have a submenu of a given menu open, and
implements them where required in the current codebase. Should hopefully
make future menu updating/syncing a bit cleaner.