Implement bridge, plug and bag editors + bridge and plug debug info (#155)
Adds bridge, plug, and bag editors to the structure editor + menu, along
with debug info toggles for both the bridge and plug editors - doesn't
seem to be much point in one for the bags.

Doesn't add presets for these yet, but should be fairly straightforward
to do based on the gate presets.

Plugs are currently a bit iffy - they usually are killed by the manager
rather than just hidden, so the models get a bit weird. I've also had to
implement some logic to re-fill the water boxes around the plugs, seems
to work fairly nicely, but there's a slight delay. I've made it so the
menu closes automatically on updating plug status, so people can't spawn
10 before they resolve, but I'm not entirely convinced it won't crash if
someone really wants it to. I think that's a later problem to resolve,
having the functionality seems more important at the moment, but fixes
welcome - otherwise, I'll open an issue for it once this is merged.

Also, plugs don't automatically update their associated hole effects or
toggles yet either, so technically you could kill the SR plug, respawn
it, and then enter the cave with the plug still alive. Again, bug issue
for later, will open issue on merge.

---------

Co-authored-by: mayabyte <mayahayden@protonmail.com>