Fix: dismiss positions no longer include pikmin on plants and grass (#181)
Adds an `is_working` inline, for ignoring pikmin that might have the
active navi as their current navi, but who shouldn't be included in the
"squad" for dismiss positions and freecam warping, etc. Also adds
ACT_Crop (plants) and ACT_Weed (nectar grass) to the list of pikmin
actions we should ignore.