Make crash landing cutscene skippable (#185)
Crash landing cutscene is now skippable after the Valley of Repose text
disappears, with an offset added to the timer automatically based on how
long you waited before skipping. The timer cutscene skipping
functionality has also been reworked slightly to allow for more
cutscenes to be skipped and offset in the future. I believe this also
fixes a bug to do with skipping treasure cutscenes after they hit the
ship/pod not updating the timer appropriately.

Due to the reworked functionality, some cutscene length offsets may need
adjusting, but that should be straightforward to tweak down the track.

The skipping cutscenes option is now the same toggle for all cutscenes
(treasures and intro so far), which I think is fair, but open to input
on that.