Respawn treasures and enemies AG (#168)
- Treasures and enemies can now be respawned via presets
- PoD presets have been adjusted with mostly correct enemy and treasure
spawns. There are some small discrepancies (no way to specify a position
for Ursidae on enter FC for example) but those can be addressed in a
future PR as this one is already way too big.
- Adds basic generator debug info similar to enemy debug info. Currently
only shows name and position but can be expanded in the future.
- Fixes an issue with treasure moving where they'd be placed halfway in
the ground.
- Sync treasure menu upon opening
- Replace menu images for pikmin colors with 32x32 variants