Rework enter method option for warping (#183)
Makes it so the "enter method" option for the warp menu only appears
when no preset is selected ("none"). Otherwise, presets will use a
pre-assigned enter method when warping to reflect what happens in the
preset.

This was meant to fix issue #169 , but this solution still does not
work. Notably, using the EC or COS presets still cause pikmin to fall
from the sky when warping, despite them being set to have a "from map
screen" enter option. Additionally, this does not fix the issue if
warping with "none" selected.

Additionally, a minor known issue is that since the COS preset is
deisgned for the cave itself, it considers the "first area enter"
cutscene for perplexing pool already played, making enter practice less
accurate.

---------

Co-authored-by: Dean Southwood <deancsouthwood@gmail.com>