Basic skeleton for integration tests (#164)
Adds an extremely basic and feature-incomplete framework for integration
tests. Tests take the form of a list of TestOps to be performed. TestOps
can be single button presses, waiting for an arbitrary state to be
reached, or in the future directly editing the game or menu state. There
are some simple combinator TestOps to do more complex actions. Macros
are provided to make writing tests (sequences of TestOps) simpler.

Test code is all behind conditional compilation `#ifdef`s. Pass the
`--test` flag when building (e.g. `python3 build.py --test`) to build
with testing code enabled. Building without the test flag should have
zero (or extremely minimal) effect on binary size. Tests will begin
running as soon as GZ is inited, which is currently upon entering file
select. Sequencing info is printed to the log via OSReports.

This is meant to be an _extremely basic first pass_ at how the
integration tests should look. When we want to write feature tests, we
will inevitably need to add more kinds of TestOps and more macros. I
want to put this out there and possibly merge it in this early state to
gather feedback and let everyone contribute to how this system looks and
works as we write tests.

**I want three reviews on this PR before merging** to ensure we're all
on the same page about this system.