Move treasures back to their original positions when retrying AGs (#208)
Previously when warping to an AG segment while some of its force-spawned
treasures are alive, they would be left in the same position they were
in when warping instead of respawned like they should be.

Also adds support for position overrides for treasure spawns. It's not
used yet since we don't have a way to get good coordinates in-game, but
it will be useful for AT splits and Enter FC.