Add option to choose active captain when warping to caves (#217)
Adds an option to choose which captain you start as when warping to
caves.

It updates according to what captain you would normally enter the next
sublevel as. If you're playing as Olimar but you last entered a hole
with Louie, it will read Louie when opening the menu.

The menu doesn't show up when warping to AGs since you always start as
Olimar in AGs.