Fix: drastically cut DOL size by un-inlining iterator functions (#227)
Cuts the size of the DOL by a fair chunk (without changing or removing
any functionality) by forcing `Iterator` functions to not inline. This
creates about an extra ~110KB of space in the game heap, which we could
siphon to the system heap for our purposes if we want to down the line.

Credit to @PikHacker for the idea - thank you!