Fix crash caused by deallocating a shared Preset (#232)
We now share presets between segments on retry, but when the ring buffer
of segments wrapped around it was still trying to free the preset owned
by the segment. This caused a use-after-free when retrying the same
sublevel 33 times. This has been fixed by correcting some of the
RingBuffer logic and by not freeing presets when destructing Segments. A
memory leak is technically introduced by this PR, but it's not one that
should manifest unless a user plays an extremely long session.