Support sprouts in presets (#224)
Adds support for sprouts in Presets. 

Sprouts can be defined either as single fixed-location seeds, or as a
number of seeds to spawn randomly distributed around a circle as if
ejected by an onion. See comments on `gz::Preset::Sprout` for how to do
either.

This PR doesn't add sprouts to presets yet.