Read Presets from files (#249)
Reads presets from files to save on memory and binary size. 

- Each preset is to be defined in a separate file. 
- At build time, the `files/presets` folder is read and an
`all_presets.txt` file is generated with a list of all preset files.
This serves as the master list of presets to load.
- At init time, all the preset files are read and a `PresetPreview` is
constructed with just the name, category, and squad info so it can be
shown in the menu.
- When the preset actually needs to be used for warping, the full thing
is read from file at that point.
- Presets have been given a ref count so they can be freed at the
appropriate time even when being shared between Segments and the warp
menu, etc.
- A future improvement to save more memory could be to always deallocate
presets that originate from file since they can always be retrieved
again easily. This would require swapping between presets and preset
previews dynamically in segments, so it's not a trivial change.
- Structure overrides now have a dedicated struct and are defined by
area and position, with a new data field for defining things like number
of stages remaining on gates and bridges.
- Presets now have an origin defined based on where they came from:
file, memcard (not yet implemented but I'm aspirational), or generated.

ALL presets have been ported to files, including the commented out AT
presets.